+++ lib/grape/validations/validators/allow_blank.rb
@@ -21,7 +21,7 @@ module Grape
 
         return unless should_validate
 
-        unless value.present?
+        unless value == false || value.present?
           fail Grape::Exceptions::Validation, params: [@scope.full_name(attr_name)], message_key: :blank
         end
       end